draw snake

+++ SnakeGame/Sources/Square.swift
@@ -1,4 +1,4 @@
 struct Square {
-    let x: Int
-    let y: Int
+    let x: Double
+    let y: Double
 }
